requirement.txt
langchain_openai
langchain_core
python-dotenv
fastapi
uvicorn
pydantic
pillow
pytesseract
PyMuPDF
faiss-cpu
sentence-transformers
scikit-learn
langchain
openai
transformers
torch
google-generativeai
transformers
qdrant-client
pypdf
numpy